requirements.txt
accelerate==0.33.0
# advertorch==0.2.3
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beartype==0.18.5
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
bokeh==3.7.2
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.0.1
colorcet==3.1.0
CoLT5-attention==0.11.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
datashader==0.18.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
executing==2.1.0
facenet-pytorch==2.5.2
fastjsonschema==2.20.0
filelock==3.16.1
flake8==7.1.1
fonttools==4.54.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h11==0.14.0
holoviews==1.20.2
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
kagglehub==0.3.5
kiwisolver==1.4.7
kornia==0.7.1
lazy_loader==0.4
lcapt @ git+https://github.com/lanl/lca-pytorch.git@6811a53ea9703f98fd85c32fc415942e1b56d6b4
libcst==1.1.0
lightning-utilities==0.11.8
linkify-it-py==2.0.3
lit==18.1.8
llvmlite==0.44.0
local-attention==1.9.15
loguru==0.7.2
lpips==0.1.4
# mair @ git+https://github.com/Harry24k/MAIR.git@fcf0abae1c901829e192d856b6b3992a7b16d90d
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
# opacus==1.5.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.2
param==2.2.0
parso==0.8.4
pexpect==4.9.0
pick==2.4.0
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
portalocker==2.10.1
progress==1.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycodestyle==2.12.1
pycparser==2.22
pyct==0.5.0
pydrive==1.3.1
pyflakes==3.2.0
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytorch-fid==0.3.0
pytz==2024.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.21.0
rtpt==0.0.4
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.2
toolz==1.0.0
# torch==2.1.2+cu121
TorchFix==0.6.0
torchmetrics==1.5.2
# torchvision==0.16.2+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==2.1.0
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.3
vector-quantize-pytorch==1.17.1
wandb==0.18.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xarray==2025.3.1
xxhash==3.5.0
xyzservices==2025.1.0
yacs==0.1.8
yapf==0.40.2
yarl==1.18.0
zipp==3.20.2
